Move ValidatedFieldElement out of index
